requirements.txt
Django==3.1.3
django-allauth==0.44.0
Pillow==7.0.0
django-cleanup==5.1.0
Pillow-8.0.1
